requirements.txt
Flask==2.0.3
werkzeug==2.0.3
flask-wtf==1.0.0
wtforms==3.0.1
openai~=1.61.1
python-dotenv~=1.0.1
pdfplumber~=0.11.0
gunicorn
python-docx